# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Sentinel AML API",
        "description": "Transaction monitoring: ingestion, rule-based detection, alerts and case management. Roles: ANALYST (queue & disposition), SUPERVISOR (+ full PII, case closure, reopen), ADMIN (+ ingestion, rule/watchlist/FX configuration).",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "basic": []
        }
    ],
...
    "paths": {
...
        "/api/v1/cases": {
            "get": {
                "tags": [
                    "Cases"
                ],
                "summary": "Cases, most recently updated first",
                "operationId": "list_2",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "OPEN",
                                    "INVESTIGATING",
                                    "SAR_FILED",
                                    "CLOSED"
                                ]
                            }
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "maximum": 200,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseAmlCase"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cases"
                ],
                "summary": "Open a case from one or more alerts of the same customer (alerts become ESCALATED)",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cases/{id}/transition": {
            "post": {
                "tags": [
                    "Cases"
                ],
                "summary": "Move a case: INVESTIGATING | SAR_FILED (SUPERVISOR) | CLOSED (SUPERVISOR; closes escalated alerts)",
                "description": "409 for a transition not allowed from the current status, 403 if the role is insufficient",
                "operationId": "transition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transition"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cases/{id}/notes": {
            "post": {
                "tags": [
                    "Cases"
                ],
                "operationId": "addNote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewNote"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseNote"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cases/{id}/assign": {
            "post": {
                "tags": [
                    "Cases"
                ],
                "operationId": "assign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assign"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/rules/{code}/history": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "Every past version of a rule's configuration",
                "operationId": "history",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RuleConfigHistory"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/cases/{id}": {
            "get": {
                "tags": [
                    "Cases"
                ],
                "operationId": "get_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts/{id}/history": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Immutable audit trail of this alert (creation, aggregation, every state transition)",
                "operationId": "history_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuditLog"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateCase": {
                "required": [
                    "alertIds",
                    "reason"
                ],
                "type": "object",
                "properties": {
                    "alertIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "title": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "reason": {
                        "maxLength": 1000,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "AlertSummary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "alertRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerRef": {
                        "type": "string"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerRiskRating": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "riskScore": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "riskBand": {
                        "type": "string"
                    },
                    "ruleCodes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "hitCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "windowStart": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "windowEnd": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "assignee": {
                        "type": "string"
                    },
                    "caseId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AmlCase": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "caseRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "INVESTIGATING",
                            "SAR_FILED",
                            "CLOSED"
                        ]
                    },
                    "priority": {
                        "type": "string"
                    },
                    "assignee": {
                        "type": "string"
                    },
                    "outcome": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "rowVersion": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "AuditLog": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "entityType": {
                        "type": "string"
                    },
                    "entityId": {
                        "type": "string"
                    },
                    "action": {
                        "type": "string"
                    },
                    "fromState": {
                        "type": "string"
                    },
                    "toState": {
                        "type": "string"
                    },
                    "actor": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CaseDetail": {
                "type": "object",
                "properties": {
                    "caseInfo": {
                        "$ref": "#/components/schemas/AmlCase"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerView"
                    },
                    "alerts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertSummary"
                        }
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CaseNote"
                        }
                    },
                    "history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditLog"
                        }
                    }
                }
            },
            "CaseNote": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "caseId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "author": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CustomerView": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "externalRef": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "nationalId": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerType": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "kycRiskRating": {
                        "type": "string"
                    },
                    "piiMasked": {
                        "type": "boolean"
                    }
                }
            },
            "Transition": {
                "required": [
                    "status"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "INVESTIGATING",
                            "SAR_FILED",
                            "CLOSED"
                        ]
                    },
                    "reason": {
                        "maxLength": 1000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "disposition": {
                        "type": "string",
                        "enum": [
                            "FALSE_POSITIVE",
                            "TRUE_POSITIVE",
                            "NO_FURTHER_ACTION"
                        ]
                    }
                }
            },
            "NewNote": {
                "required": [
                    "body"
                ],
                "type": "object",
                "properties": {
                    "body": {
                        "maxLength": 4000,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "Assign": {
                "required": [
                    "assignee"
                ],
                "type": "object",
                "properties": {
                    "assignee": {
                        "maxLength": 60,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
...
            "RuleConfigHistory": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "ruleCode": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "weight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "params": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    },
                    "changedBy": {
                        "type": "string"
                    },
                    "changedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "PageResponseAmlCase": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AmlCase"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}